Move functions from keys.rs to wallet.rs
