+++ aries_vcx_core/src/indy/wallet.rs
@@ -1,10 +1,6 @@
-use crate::secret;
+use crate::{secret, utils};
 use serde::{Deserialize, Serialize};
-use vdrtools::{
-    types::domain::wallet::{default_key_derivation_method, KeyDerivationMethod},
-    types::errors::IndyErrorKind,
-    Locator,
-};
+use vdrtools::{DidMethod, DidValue, KeyInfo, Locator, MyDidInfo, types::domain::wallet::{default_key_derivation_method, KeyDerivationMethod}, types::errors::IndyErrorKind};
 
 use crate::{
     errors::error::{AriesVcxCoreError, AriesVcxCoreErrorKind, VcxCoreResult},
@@ -12,6 +8,7 @@ use crate::{
 };
 use crate::{global::settings, WalletHandle};
 use crate::{indy::keys, SearchHandle};
+use crate::indy::utils::mocks::did_mocks::{did_mocks_enabled, DidMocks};
 
 #[derive(Clone, Debug, Default, Builder, Serialize, Deserialize)]
 #[builder(setter(into, strip_option), default)]
@@ -514,7 +511,7 @@ pub async fn wallet_configure_issuer(
     enterprise_seed: &str,
 ) -> VcxCoreResult<IssuerConfig> {
     let (institution_did, _institution_verkey) =
-        keys::create_and_store_my_did(wallet_handle, Some(enterprise_seed), None).await?;
+        create_and_store_my_did(wallet_handle, Some(enterprise_seed), None).await?;
 
     Ok(IssuerConfig { institution_did })
 }
@@ -581,4 +578,76 @@ pub async fn close_wallet(wallet_handle: WalletHandle) -> VcxCoreResult<()> {
     Locator::instance().wallet_controller.close(wallet_handle).await?;
 
     Ok(())
+}
+
+pub async fn create_and_store_my_did(
+    wallet_handle: WalletHandle,
+    seed: Option<&str>,
+    method_name: Option<&str>,
+) -> VcxCoreResult<(String, String)> {
+    trace!(
+        "create_and_store_my_did >>> seed: {:?}, method_name: {:?}",
+        seed,
+        method_name
+    );
+
+    if settings::indy_mocks_enabled() {
+        return Ok((utils::constants::DID.to_string(), utils::constants::VERKEY.to_string()));
+    }
+
+    let res = Locator::instance()
+        .did_controller
+        .create_and_store_my_did(
+            wallet_handle,
+            MyDidInfo {
+                method_name: method_name.map(|m| DidMethod(m.into())),
+                seed: seed.map(ToOwned::to_owned),
+                ..MyDidInfo::default()
+            },
+        )
+        .await?;
+
+    Ok(res)
+}
+
+pub async fn libindy_replace_keys_start(wallet_handle: WalletHandle, did: &str) -> VcxCoreResult<String> {
+    if DidMocks::has_did_mock_responses() {
+        warn!("libindy_replace_keys_start >> retrieving did mock response");
+        return Ok(DidMocks::get_next_did_response());
+    }
+
+    let res = Locator::instance()
+        .did_controller
+        .replace_keys_start(wallet_handle, KeyInfo::default(), DidValue(did.into()))
+        .await?;
+
+    Ok(res)
+}
+
+pub async fn libindy_replace_keys_apply(wallet_handle: WalletHandle, did: &str) -> VcxCoreResult<()> {
+    if did_mocks_enabled() {
+        warn!("libindy_replace_keys_apply >> retrieving did mock response");
+        return Ok(());
+    }
+
+    Locator::instance()
+        .did_controller
+        .replace_keys_apply(wallet_handle, DidValue(did.into()))
+        .await?;
+
+    Ok(())
+}
+
+pub async fn get_verkey_from_wallet(wallet_handle: WalletHandle, did: &str) -> VcxCoreResult<String> {
+    if DidMocks::has_did_mock_responses() {
+        warn!("get_verkey_from_wallet >> retrieving did mock response");
+        return Ok(DidMocks::get_next_did_response());
+    }
+
+    let res = Locator::instance()
+        .did_controller
+        .key_for_local_did(wallet_handle, DidValue(did.into()))
+        .await?;
+
+    Ok(res)
 }
